requirements.txt
bokeh==1.0.4
code128==0.3
Django==2.1.5
django-formtools==2.1
django-rest-framework==0.1.0
djangorestframework==3.9.1
gunicorn==19.9.0
holoviews==1.11.2
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.1
packaging==19.0
param==1.8.2
Pillow==5.4.1
psycopg2==2.7.7
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
pyviz-comms==0.7.0
PyYAML==3.13
six==1.12.0
tornado==5.1.1
